Add box resolution code
Sequelize is slightly annoying in design. I also am going to need to do a
security review of their code base, I saw string concatenation in a commit.

At this point the JSON includes everything but the provider details and the
URLs that should be built as a result. Thats my next step. Should have a
working system shortly.